Deepening the semantic labelling of the subsections
The subsections with icons now are given meaningful names - which permits easier navigation and recognition.  For instance, the disclosure provisions permit the Recipient to disclose on three conditions, one of which is that that person is "bound" to keep the information confidential.   (Note the deep link to #Disclose.MayIf.Bound.sec
http://www.commonaccord.org/index.php?action=doc&file=G/IACCM-NDA-Design/Demo/Acme_Quake.md#Disclose.MayIf.Bound.sec